% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSelect_Star.R
\name{dbSelect_Star}
\alias{dbSelect_Star}
\title{Select Star

This functions selects stars from a table given a connection.}
\usage{
dbSelect_Star(con, db_tbl)
}
\arguments{
\item{con}{A connection object}

\item{db_tbl}{string - name of table, in schema.table format.s}
}
\value{
a dataframe/tibble with the results of the query.
}
\description{
Select Star

This functions selects stars from a table given a connection.
}
\examples{
\dontrun{

db_con <- RPostgres::dbConnect()
df <- dbSelect_Star(db_con, "schema.table")
}

}
