% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_name_dot_csv.R
\name{df_name_dot_csv}
\alias{df_name_dot_csv}
\title{Dataframe to CSV}
\usage{
df_name_dot_csv(x)
}
\arguments{
\item{x}{dataframe or tibble}
}
\value{
Returns the name of the dataframe as a text.
}
\description{
Dataframe to CSV
}
\details{
This function takes a dataframe and extracts the name of the dataframe
as text or string.  You can then use this string to name a csv file.
}
\examples{
df <- as.data.frame(datasets::cars)
df_name_dot_csv(df)

}
