% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_mean.R
\name{geom_mean}
\alias{geom_mean}
\title{Geometric Mean
This functions creates a geometric mean of a numeric vector.}
\usage{
geom_mean(x, roll_window = length(x))
}
\arguments{
\item{x}{numeric vector or dataframe column.}

\item{roll_window}{numeric - length of the vector, or can be used to build a
rolling geometric mean.}
}
\value{
Returns the geometric mean of a vector, or one number.
}
\description{
Geometric Mean
This functions creates a geometric mean of a numeric vector.
}
\examples{
library(tibbletime)
library(purrr)
geom_mean_3 <- purrr::partial(geom_mean, roll_window = 3)
geom_mean_roll_3 <- tibbletime::rollify(geom_mean_3,window = 3)
data <- c(20,25,38,15,70)
geom_mean(data)
geom_mean_roll_3(data)

}
