% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pkgs_user_lib.R
\name{install.pkgs.user_lib}
\alias{install.pkgs.user_lib}
\title{Install Packages into R_LIBS_USER}
\usage{
install.pkgs.user_lib(pkgs, ...)
}
\arguments{
\item{pkgs}{string character vector list of packages}

\item{...}{any other arguments passed to install.packages()}
}
\description{
This function will install into the folder listed under the environmental
variable R_LIBS_USER. The user can specify this in the systems environmental
variables or in the .Renviron file.
}
\details{
Installs packages into the folder referenced by the R_LIBS_USER environmental
variable.
}
\examples{
\dontrun{
  pkgs <- c("RODBC","RJDBC")
  install.pkgs.user_lib(pkgs = pkgs, dependencies = TRUE)
}

}
