% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgCreatePrimaryKey.R
\name{pgCreatePrimaryKey}
\alias{pgCreatePrimaryKey}
\title{Create Primary Keys for Postgres Table}
\usage{
pgCreatePrimaryKey(pg_conn, schema_name, table_name, primary_keys = "NA")
}
\arguments{
\item{pg_conn}{connection string Postgres connection pointer/string (PqConnection)
created from RPostgres::Postgres(). JDBC connections will break this function.}

\item{schema_name}{character Posgres schema name.}

\item{table_name}{character Postgres table name.}

\item{primary_keys}{character list of primary keys.}
}
\value{
The results will look like the print out below.  It is a generic print
out from RPostgres.  Check the table metadata in Postgres server to check primary
key build.
}
\description{
Description
}
\details{
The table name needs to include the schema in the Postgres server.
The list of primary keys can simply be "var1,var2,var3" because this will be
inserted into an SQL statement that will be sent to the server.
}
