% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgCreateTempTable.R
\name{pgCreateTempTable}
\alias{pgCreateTempTable}
\title{Create Primary Indexes for Postgres Table}
\usage{
pgCreateTempTable(pg_conn, schema_tbl, index_cols, query)
}
\arguments{
\item{pg_conn}{connection string Postgres connection pointer/string (PqConnection)}

\item{schema_tbl}{character Schema.table name.}

\item{index_cols}{character Expects a vector of column names...c('col1','col2','col3')}

\item{query}{character Query string to run in database.
created from RPostgres::Postgres(). JDBC connections will break this function.}
}
\value{
The results will look like the print out below.  It is a generic print
out from RPostgres.  Check the table metadata in Postgres server to check primary
key build.
}
\description{
Description
}
\details{
The table name needs to include the schema in the Postgres server.
The list of primary keys can simply be "var1,var2,var3" because this will be
inserted into an SQL statement that will be sent to the server.
}
