% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad2deg.R
\name{rad2deg}
\alias{rad2deg}
\title{Convert radians to degree}
\usage{
rad2deg(r)
}
\arguments{
\item{r}{numeric - radians, or output from a trigonmetric functions.}
}
\value{
A number in degrees that was converted from radians.
}
\description{
Convert radians to degree: d = r x 180 x (1/pi)
}
\examples{
rads <- sin(4/5)
rad2deg(rads)

}
