% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_db_name.R
\name{rm_db_name}
\alias{rm_db_name}
\title{Remove Database Names from Columns Names.}
\usage{
rm_db_name(df, db_name)
}
\arguments{
\item{df}{Dataframe}

\item{db_name}{The name of the table in the database that you want removed from the column names.}
}
\description{
Remove the database column extensions that are created when query databases such as Hive.
}
