% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RToolShed-package.R
\docType{package}
\name{RTooShed}
\alias{RToolShed}
\alias{RToolShed-package}
\alias{RTooShed-package}
\title{RToolShed: Tools for Data Analysis and Data Science}
\description{
There are times as a data analyst or a data scientist that we find ourselves writing
    the same code over and over across projects.  The central idea of the R Tool Shed package is
    to centralize the code to improve development and to standardize the functionality of the 
    functions across projects.
}
\author{
Alfredo G Marquez
}
\keyword{internal}
