% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_na.R
\name{test_na}
\alias{test_na}
\title{Test if a Column/Variable has NA's}
\usage{
test_na(.data, col_n)
}
\arguments{
\item{.data}{dataframe or tibble}

\item{col_n}{string name of the column/variable within the dataframe.}
}
\value{
Outputs a message of whether a column has NA's.
}
\description{
Test if a Column/Variable has NA's
}
\details{
Sums up the total number of NA's and reports where a column has 1 or more NA's
}
\examples{
\dontrun{
library(magrittr)
library(dplyr)
library(glue)
library(testthat)
test_na(mtcars,'cyl')
}

}
