% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_postgres.R
\name{to_postgres}
\alias{to_postgres}
\title{Push R Dataframes to Postgres shemas other than Public.}
\usage{
to_postgres(df, pg_conn, table_name, schema_name, orderby = "NA",
  primary_keys = "NA")
}
\arguments{
\item{df}{Dataframe that will be pushed to Postgres}

\item{pg_conn}{Connection String}

\item{table_name}{Name of new table in Postgres}

\item{schema_name}{Name of final schema if not public, default is public}

\item{orderby}{Fields to order by}

\item{primary_keys}{list of primary keys}
}
\description{
Push R Dataframes to Postgres shemas other than Public.
}
\details{
#' require(RPostgres)
require(stringr)
require(dbplyr)package parameters:

df <- brck_mrtr_dim  #1 input table
pg_conn <- db_postgres #2 pg conn
table_name <- "sister_app_logs" #3 name of new table
schema_name <- "landing" #4 name of final schema if not public
orderby <- "NA" #"co_loc_ref_I"
primary_keys <- "NA" # paste("co_loc_i", "co_loc_ref_i", sep = ",")#5 list of primary keyspackage parameters:
}
\examples{
## NOT RUN ##
## to_postgres(df = DF, pg_conn = db_CON, table_name = "data_upload", schema_name = "prod", orderby = "group", primary_keys = "group,date") ##
}
