% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csv_to_hive.R
\name{write_csv_to_hive}
\alias{write_csv_to_hive}
\title{Write a CSV file to Hive}
\usage{
write_csv_to_hive(csv_file, id, server, schema_table)
}
\arguments{
\item{csv_file}{CSV file to upload; will be converted to dataframe by dplyr::read_csv.}

\item{id}{string ID of user}

\item{server}{string server extention or path}

\item{schema_table}{string "schema.table" Name of the table to write to in Hive.}
}
\value{
Does not return anything.
}
\description{
Write a CSV file to Hive
}
\details{
Uploads a CSV file and uploads it to Hive.  This assumes that when you log into
Hive/Hadoop, the login is similar to `XXXXX@edge.hadoop.co.com``
}
\examples{
\dontrun{
library(ssh)
library(dplyr)
library(readr)
library(rstudioapi)
library(magrittr)
df <- mtcars
zid <- 'XXXXX'
server <- 'edge.hadoop.co.com'
schema_table <- 'schema.table'
.pwd <- rstudioapi::askForPassword('password')
file <- c('table_for_hive.csv')
write_csv_to_hive(csv_file = file, id = zid, server = server, schema_table = schema_table)
}

}
