% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_df_to_hive3.R
\name{write_df_to_hive3}
\alias{write_df_to_hive3}
\title{Write a Dataframe to Hive 3}
\usage{
write_df_to_hive3(df, id, schema, table, server, append_data = FALSE)
}
\arguments{
\item{df}{must be a dataframe/tibble}

\item{id}{string ID of user. Password will be requested at function call.}

\item{schema}{string schema name in hive}

\item{table}{string name for the table in hive.  One will be created if not
exists.}

\item{server}{string server extention or path}

\item{append_data}{logical, defaults to FALSE for overwrite; TRUE appends the
to the data.}
}
\value{
Does not return anything.
}
\description{
Write a Dataframe to Hive 3
}
\details{
Uploads an R dataframe to the edge node as a CSV, uploads it to
Hive, and creates a managed table. The function also cleans up the csv file
on the edge node and in the users' hdfs home location.  This assumes that
when you log into Hive/Hadoop, the login is similar to
\code{XXXXX@edge.hadoop.co.com}
}
\examples{
\dontrun{
library(ssh)
library(dplyr)
library(readr)
library(askpass)
library(magrittr)

df <- as.data.frame(Seatbelts)
id <- 'XXXXX'
server <- 'edge.hadoop.co.com'
schema <- 'schema'
table <- 'table'
write_df_to_hive3(df =
                   id = id,
                   schema = schema,
                   table = table,
                   server = server,
                   append_data = FALSE)
}

}
