% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_it.R
\name{diff_it}
\alias{diff_it}
\title{Difference Variables \code{n} Lags, \code{k} Differences}
\usage{
diff_it(df, col = NULL, lags = 1, diff = 1, mutate_type = c("mutate", "trans"))
}
\arguments{
\item{df}{dataframe or tibble}

\item{col}{string - specify the column/field/variable for differencing.}

\item{lags}{number - specify the number of lags for differencing.}

\item{diff}{number - specify the number of differences for differencing.}

\item{mutate_type}{string - select whether to \code{mutate()} or \code{transmute()} from dplyr.}
}
\value{
A dataframe with the lagged differences of the column specified.
}
\description{
This function takes a column and builds the differences based
on the difference \code{k} between the column and the lag \code{n} specified.
}
\details{
Build the differences of a variable
}
\examples{
library(magrittr)
df <- datasets::airmiles \%>\% as.vector() \%>\% tibble::enframe(name = "airmiles")
diff_it(df, col = "airmiles") \%>\% head()

}
