% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_helper_functions.R
\name{SQL_HELPER_FUNCTIONS}
\alias{SQL_HELPER_FUNCTIONS}
\alias{select_star}
\alias{drop_if_exists}
\alias{add_verbose}
\alias{create_index}
\title{Helper functions for common SQL Statements}
\usage{
select_star(schema_tbl)

drop_if_exists(schema_tbl)

add_verbose(schema_tbl)

create_index(schema_tbl, index_cols)
}
\arguments{
\item{schema_tbl}{character Schema.table name.}

\item{index_cols}{character expects a vector of column names...c('col1','col2','col3')}
}
\value{
These functions return common SQL statements that can be combined to
to build a complete pipeline of commands such as drop, create, create index.
}
\description{
Helper functions for common SQL Statements
}
\details{
The table name needs to include the schema in the Postgres server.
The list of primary keys can simply be "var1,var2,var3" because this will be
inserted into an SQL statement that will be sent to the server.
}
