% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_postgres.R
\name{to_postgres}
\alias{to_postgres}
\title{Push R Dataframes to Postgres schemas other than Public.}
\usage{
to_postgres(df, pg_conn, schema_name = "public", table_name,
  primary_keys = "NA")
}
\arguments{
\item{df}{Dataframe that will be pushed to Postgres, required.}

\item{pg_conn}{<required> Postgres connection (PqConnection) created from RPostgres::Postgres().
JDBC connections will break this function.}

\item{schema_name}{<string> Name of final schema if not public, default is public.}

\item{table_name}{<string> Name of new table in Postgres, required.}

\item{primary_keys}{<string< Fields to for primary keys, used in a SQL statement, default is NA.}
}
\value{
Paste copy of output here.  You will get a generic message from RPostgres, check
Postgres server for the data to ensure that the data was transferred.
}
\description{
There some cases where the Postgres servers are set up so that R is not able to build tables
in schemas other than public. This function will delete any named tables that exit in both
the public schema and the destination schema, then it will re-build the data into public then
into the desired schema.
}
\details{
No Detailspackage parameters:
}
\examples{
\dontrun{
to_postgres(df = DF, pg_conn = db_CON, table_name = "data_upload", schema_name = "prod",
primary_keys = "group,date")
}

}
