% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_df_to_hive.R
\name{write_df_to_hive}
\alias{write_df_to_hive}
\title{Write an R dataframe to Hive}
\usage{
write_df_to_hive(df, id, server, schema_table)
}
\arguments{
\item{df}{<dataframe> Dataframe to upload.}

\item{id}{<string> ID of user}

\item{server}{<string> server extention or path}

\item{schema_table}{<string> "schema.table" Name of the table to write to in Hive.}
}
\value{
Does not return anything.
}
\description{
Write an R dataframe to Hive
}
\details{
Uploads an R dataframe and uploads it to Hive.  This assumes that when you log into
Hive/Hadoop, the login is similar to `XXXXX@edge.hadoop.co.com``
}
\examples{
\dontrun{
library(ssh)
library(dplyr)
library(readr)
library(rstudioapi)
df <- mtcars
zid <- 'XXXXX'
server <- 'edge.hadoop.co.com'
schema_table <- 'schema.table'
.pwd <- rstudioapi::askForPassword('password')
write_df_to_hive(df = df, id = zid, server = server, schema_table = schema_table)
}

}
